SAdemo/code/requirements.txt
albumentations==1.4.4
annotated-types==0.6.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
Flask==3.0.3
fonttools==4.51.0
fsspec==2023.4.0
idna==3.7
imageio==2.34.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.23.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile
torch
torchaudio
torchvision
tqdm
typing_extensions
tzdata
ultralytics==8.2.1
urllib3==2.2.1
Werkzeug==3.0.2
gunicorn
